pyproject.toml
[project]
name = "consciousness-experiment"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "langfuse>=4.0.1",
    "python-dotenv>=1.2.2",
]
